+++ examples/unfolded/demo_vanilla_unfolded.py
@@ -68,7 +68,7 @@ num_workers = 4 if torch.cuda.is_available() else 0
 
 # Degradation parameters
 factor = 2
-noise_level_img = 0.03
+noise_level_img = 0.0075
 
 # Generate the gaussian blur downsampling operator.
 physics = dinv.physics.Downsampling(
@@ -104,6 +104,10 @@ test_dataset = dinv.datasets.HDF5Dataset(path=generated_datasets_path, train=Fal
 # Note that if the prior (resp. a parameter) is initialized with a list of length max_iter,
 # then a distinct model (resp. parameter) is trained for each iteration.
 # For fixed trained model prior (resp. parameter) across iterations, initialize with a single element.
+#
+# .. tip::
+#          While in theory any initial stepsize could be adapted during training, in practice the best performance
+#          is obtained by starting with a stepsize close to the largest stable one, i.e. :math:`1/\|A\|^2`
 
 # Unrolled optimization algorithm parameters
 max_iter = 5  # number of unfolded layers
@@ -116,7 +120,10 @@ data_fidelity = L2()
 prior = PnP(denoiser=dinv.models.DnCNN(depth=20, pretrained="download").to(device))
 
 # The parameters are initialized with a list of length max_iter, so that a distinct parameter is trained for each iteration.
-stepsize = [1.0] * max_iter  # stepsize of the algorithm
+norm = physics.compute_norm(
+    torch.ones((1, 3, img_size, img_size), device=device), squared=True
+)
+stepsize = [1.0 / norm] * max_iter  # stepsize of the algorithm
 sigma_denoiser = [
     1.0
 ] * max_iter  # noise level parameter of the denoiser (not used by DnCNN)
@@ -146,10 +153,13 @@ model = DRS(
 # Define the training parameters.
 # ----------------------------------------------------------------------------------------
 # We use the Adam optimizer and the StepLR scheduler.
-
+#
+# .. note::
+#       We use a pretrained checkpoint here (trained for 30 epochs) to speed up the demo,
+#       but you can reproduce results by training for the whole 30 epochs from scratch.
 
 # training parameters
-epochs = 5 if torch.cuda.is_available() else 1
+epochs = 3 if torch.cuda.is_available() else 1
 learning_rate = 5e-4
 train_batch_size = 32 if torch.cuda.is_available() else 1
 test_batch_size = 3
@@ -157,15 +167,15 @@ test_batch_size = 3
 # choose optimizer and scheduler
 optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate, weight_decay=1e-8)
 
-# If working on CPU, start with a pretrained model to reduce training time
-if not torch.cuda.is_available():
-    file_name = "demo_vanilla_unfolded.pth"
-    url = get_weights_url(model_name="demo", file_name=file_name)
-    ckpt = torch.hub.load_state_dict_from_url(
-        url, map_location=lambda storage, loc: storage, file_name=file_name
-    )
-    model.load_state_dict(ckpt["state_dict"])
-    optimizer.load_state_dict(ckpt["optimizer"])
+# start with a pretrained model to reduce training time
+# comment out if you wan't to train from scratch
+file_name = "demo_vanilla_unfolded.pth"
+url = get_weights_url(model_name="demo", file_name=file_name)
+ckpt = torch.hub.load_state_dict_from_url(
+    url, map_location=lambda storage, loc: storage, file_name=file_name
+)
+model.load_state_dict(ckpt["state_dict"])
+optimizer.load_state_dict(ckpt["optimizer"])
 
 # choose supervised training loss
 losses = [dinv.loss.SupLoss(metric=dinv.metric.MSE())]